+++ example/content/dns-tfstate.tf
@@ -1,0 +1,1 @@
+../shared/halostatue_ca_dns-tfstate.tf